# docs/swagger.yaml
basePath: /api
definitions:
...
  domain.Plan:
    enum:
    - trial
    - base
    - professional
    - ai
    type: string
    x-enum-varnames:
    - PlanTrial
    - PlanBase
    - PlanProfessional
    - PlanAI
...
  domain.User:
    properties:
      active:
        type: boolean
      created_at:
        type: string
      email:
        type: string
      email_verified:
        type: boolean
      id:
        type: string
      name:
        type: string
      plan:
        $ref: '#/definitions/domain.Plan'
      role:
        $ref: '#/definitions/domain.UserRole'
      subscription_active:
        type: boolean
      trial_ends_at:
        type: string
    type: object
  domain.UserRole:
    enum:
    - admin
    - seller
    type: string
    x-enum-varnames:
    - RoleAdmin
    - RoleSeller
...
  httphandler.authResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      user:
        $ref: '#/definitions/domain.User'
    type: object
  httphandler.checkoutRequest:
    properties:
      cvv:
        example: "123"
        type: string
      email:
        example: joao@email.com
        type: string
      expiry_month:
        example: "12"
        type: string
      expiry_year:
        example: "2030"
        type: string
      holder_name:
        example: João Silva
        type: string
      number:
        example: "4444444444444444"
        type: string
      plan:
        example: professional
        type: string
    required:
    - cvv
    - email
    - expiry_month
    - expiry_year
    - holder_name
    - number
    - plan
    type: object
...
  httphandler.errorResponse:
    properties:
      code:
        example: email_not_verified
        type: string
      error:
        example: mensagem de erro
        type: string
    type: object
  httphandler.forgotPasswordRequest:
    properties:
      email:
        example: joao@email.com
        type: string
    type: object
...
  httphandler.loginRequest:
    properties:
      email:
        example: admin@rinoseller.com
        type: string
      password:
        example: "123456"
        type: string
    required:
    - email
    - password
    type: object
  httphandler.messageResponse:
    properties:
      message:
        example: operação realizada com sucesso
        type: string
    type: object
...
  httphandler.registerRequest:
    properties:
      email:
        example: joao@email.com
        type: string
      name:
        example: João Silva
        type: string
      password:
        example: "123456"
        type: string
      plan:
        example: trial
        type: string
    required:
    - email
    - name
    - password
    - plan
    type: object
  httphandler.resendVerificationRequest:
    properties:
      email:
        example: joao@email.com
        type: string
    required:
    - email
    type: object
  httphandler.resetPasswordRequest:
    properties:
      password:
        example: novaSenha123
        type: string
      token:
        example: abc123
        type: string
    type: object
...
  httphandler.verifyEmailRequest:
    properties:
      token:
        example: abc123
        type: string
    required:
    - token
    type: object
host: localhost:8080
info:
  contact: {}
  description: API do sistema RinoSeller — gestão de vendas, clientes, produtos, orçamentos
    e finanças.
  title: RinoSeller API
  version: "1.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: E-mail
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httphandler.forgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httphandler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Esqueci a senha
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Credenciais
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httphandler.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httphandler.authResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Login
      tags:
      - Auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.User'
      security:
      - BearerAuth: []
      summary: Meu perfil
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Dados de registro
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httphandler.registerRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/httphandler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Registrar conta
      tags:
      - Auth
  /auth/resend-verification:
    post:
      consumes:
      - application/json
      parameters:
      - description: E-mail
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httphandler.resendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httphandler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Reenviar e-mail de confirmação
      tags:
      - Auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Token e nova senha
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httphandler.resetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httphandler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Redefinir senha
      tags:
      - Auth
  /auth/verify-email:
    post:
      consumes:
      - application/json
      parameters:
      - description: Token de verificação
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httphandler.verifyEmailRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httphandler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Confirmar e-mail
      tags:
      - Auth
...
  /subscriptions/checkout:
    post:
      consumes:
      - application/json
      description: Simula a cobrança via Asaas e ativa a assinatura do usuário. Não
        requer login — usado tanto no fluxo de cadastro quanto a partir da tela de
        cobrança após o período de teste.
      parameters:
      - description: Dados do plano e do cartão
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/httphandler.checkoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/httphandler.messageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/httphandler.errorResponse'
      summary: Assinar um plano (checkout simulado)
      tags:
      - Subscriptions
...
securityDefinitions:
  BearerAuth:
    description: 'Informe: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"